refactor: move fillTables script code to scripts folder
